requirements.txt
termcolor
dm_control>=1.0.0
git+https://github.com/denisyarats/dmc2gym.git
tensorboard
imageio
imageio-ffmpeg
hydra-core==1.1.2
omegaconf==2.1.2
pandas
ipdb
tqdm
yapf
scikit-learn
matplotlib
kornia
scikit-image
scikit-video
opencv-python
pillow
mujoco==2.3.7
